### 2022-11-03 #116 @whedon
[PRE REVIEW]: JustSayIt.jl: A Fresh Approach to Open Source Voice Assistant Development
**Submitting author:** @omlins (<a href="http://orcid.org/0000-0001-8969-5619">Samuel Omlin</a>)
**Repository:** <a href="https://github.com/omlins/JustSayIt.jl" target ="_blank">https://github.com/omlins/JustSayIt.jl</a>
**Version:**
