### 2022-11-11 a46f354
add pretty print for sequences
