### 2022-11-11 83ff4c4
add pretty print for sequences
