+++ paper/paper.yml
@@ -10,7 +10,7 @@ keywords:
   - Voice function arguments
 authors:
   - name: Samuel Omlin
-    orcid: 0000-0000-0000-000X
+    orcid: 0000-0001-8969-5619
     affiliation: 1
 affiliations:
   - name: Swiss National Supercomputing Centre (CSCS), ETH Zurich, Lugano, Switzerland
