### 2022-12-05 8f3ccef
enable hierarchical commands
