+++ src/Commands/Keyboard.jl
@@ -16,7 +16,7 @@ module Keyboard
 
 using PyCall
 using ..Exceptions
-import ..JustSayIt: @voiceargs, pyimport_pip, controller, set_controller, PyKey, default_language, type_languages, lang_str, LANG, LANG_CODES_SHORT, LANG_STR, ALPHABET, DIGITS, MODELTYPE_DEFAULT, MODELNAME, modelname, tic, toc, is_next, are_next, all_consumed, was_partial_recognition, InsecureRecognitionException, reset_all, do_delayed_resets, interpret_enum
+import ..JustSayIt: @voiceargs, pyimport_pip, controller, set_controller, PyKey, default_language, type_languages, lang_str, LANG, LANG_CODES_SHORT, LANG_STR, ALPHABET, DIGITS, MODELTYPE_DEFAULT, MODELNAME, modelname, tic, toc, is_next, are_next, all_consumed, was_partial_recognition, InsecureRecognitionException, reset_all, do_delayed_resets, interpret_enum, UNKNOWN_TOKEN
 
 
 ## PYTHON MODULES
@@ -180,10 +180,12 @@ Type words only (upper and lower case). Supported keywords are:
 - "lowercase"
 
 #### `letters`
-Type letters only. Supported keywords are:
+Type letters only (upper and lower case). Supported keywords are:
 - "terminus"
 - "undo"
 - "redo"
+- "uppercase"
+- "lowercase"
 - "space"
 
 #### `digits`
@@ -198,7 +200,7 @@ Type digits only. Supported keywords are:
 type
 @enum TokenGroupKind undefined_kind keyword_kind word_kind letter_kind digit_kind language_kind direct_kind
 @enum TypeMode text words letters digits
-@voiceargs (mode=>(valid_input=Tuple(TYPE_MODES), interpret_function=interpret_typemode)) function type(mode::TypeMode; end_keyword::String=TYPE_END_KEYWORD, active_lang=type_languages()[1], do_keystrokes::Bool=true)
+@voiceargs (mode=>(valid_input=Tuple(TYPE_MODES), interpret_function=interpret_typemode)) function type(mode::TypeMode; end_keyword::String=TYPE_END_KEYWORD, exit_on_unknown=false, max_word_groups=Inf, active_lang=type_languages()[1], do_keystrokes::Bool=true)
     @info "Typing $(string(mode))..."
     type_memory      = Vector{String}()
     tokengroup_str   = ""
@@ -227,6 +229,7 @@ type
             tokengroup_str = ""
             ig += 1
             it = 0
+            if (ig > max_word_groups) break end
         end
         if is_new_group && (tokengroup_kind == undefined_kind)
             reset_all(; hard=true, exclude_active=true)      # NOTE: a reset is required to avoid that the dynamic recognizers generated in is_next and are_next include the previous recognition as desired for normal command recognition.
@@ -270,7 +273,7 @@ type
                     break
                 end
                 if keyword == end_keyword
-                    is_typing = false
+                        is_typing = false
                 elseif keyword == TYPE_KEYWORDS[active_lang]["undo"]
                     if ig > 1
                         ig -= 1
@@ -360,6 +363,7 @@ type
                 nb_keyword_chars += length(keyword_sign)
             end
         else
+            token = ""
             if     (tokengroup_kind == word_kind)     token = next_word(active_lang)
             elseif (tokengroup_kind == letter_kind)   token = next_letter(active_lang)
             elseif (tokengroup_kind == digit_kind)    token = next_digit(active_lang)
@@ -369,47 +373,51 @@ type
[…]
