### 2022-12-05 e92f553
enable hierarchical commands
