### 2022-12-05 3868f1a
enable hierarchical commands
