+++ src/next_token.jl
@@ -40,8 +40,8 @@ are_next
 
 let
     global next_token, is_next, _is_next, are_next, _are_next, recognizer, force_reset_previous, all_consumed, was_partial_recognition, reset_all, do_delayed_resets # NOTE: recogniser needs to be declared global here, even if elsewhere the method created here might not be used, as else we do not have access to the other reconizer methods here.
-    recognizers_to_reset = Vector{PyObject}()
-    active_recognizer::Union{Nothing, PyObject} = nothing
+    recognizers_to_reset = Vector{Recognizer}() 		      # NOTE: only persistent recognizer will need to be reset; temporary recognizers will automatically be removed by the python garbage collector (see __del__ in Vosk source).
+    active_recognizer::Union{Nothing, Recognizer} = nothing
     was_partial_result = false
     token_buffer = Vector{String}()
     i = 0
@@ -49,15 +49,17 @@ let
 	was_partial_recognition()::Bool = was_partial_result
 
 
-    function _next_token(recognizer::PyObject, noise_tokens::AbstractArray{String}; consume::Bool=true, timeout::Float64=Inf, use_partial_recognitions::Bool=false, restart_recognition::Bool=false, ignore_unknown::Bool=true)
+    function _next_token(recognizer::Recognizer, noise_tokens::AbstractArray{String}; consume::Bool=true, timeout::Float64=Inf, use_partial_recognitions::Bool=false, restart_recognition::Bool=false, ignore_unknown::Bool=true)
 		ignore_tokens = ignore_unknown ? [noise_tokens..., UNKNOWN_TOKEN] : noise_tokens
 		if (recognizer != active_recognizer && !isnothing(active_recognizer) && !isempty(token_buffer) && i==0) # If the recognizer was changed despite that tokens were recognized, but none was consumed, then we will always want to restart recognition.
 			 restart_recognition = true
 		end
 		if (!was_partial_result) do_delayed_resets(;hard=false) end                                 # When a result was found, then soft resets that were previously delayed to keep latency minimal can now be performed.
-		if (recognizer != active_recognizer && was_partial_result && !isnothing(active_recognizer)) # Reset the active recognizer if a new recognizer will become active. #NOTE: Reset after a result is not needed and a hard reset leads to the following Vosk error in that case: ASSERTION_FAILED (VoskAPI:ComputeFinalCosts():lattice-faster-decoder.cc:540) Assertion failed: (!decoding_finalized_)
-			if restart_recognition push!(recognizers_to_reset, active_recognizer)
-			else                   reset(active_recognizer; hard=true)
+		if !isnothing(active_recognizer) && active_recognizer.is_persistent
+			if (recognizer != active_recognizer && was_partial_result) # Reset the active recognizer if a new recognizer will become active. #NOTE: Reset after a result is not needed and a hard reset leads to the following Vosk error in that case: ASSERTION_FAILED (VoskAPI:ComputeFinalCosts():lattice-faster-decoder.cc:540) Assertion failed: (!decoding_finalized_)
+				if restart_recognition push!(recognizers_to_reset, active_recognizer)
+				else                   reset(active_recognizer; hard=true)
+				end
 			end
 		end
 		t  = 0.0
@@ -78,7 +80,7 @@ let
[…]
