### 2022-12-05 251c5f8
enable hierarchical commands
