### 2022-12-05 3a6816d
enable hierarchical commands
