### 2022-12-05 5fa507d
enable hierarchical commands
