### 2022-12-05 f25d230
enable hierarchical commands
