### 2022-12-06 1717c5b
enable using dynamic recognizers only
