### 2022-12-06 8077344
enable using dynamic recognizers only
