### 2022-12-08 da35496
add function to type letter
