+++ src/Commands/Keyboard.jl
@@ -453,7 +453,7 @@ function handle_uppercase(token::String, is_uppercase::Bool, active_lang::String
 end
 
 
-    "Get next word from speech."
+"Get next word from speech."
 function next_word(lang::String)
     if     (lang == LANG.DE   ) next_word_DE()
     elseif (lang == LANG.EN_US) next_word_EN_US()
@@ -469,8 +469,8 @@ end
 
 
 "Get next letter from speech."
-function next_letter(lang::String)
-    if is_next(UNKNOWN_TOKEN, [keys(ALPHABET[lang])...]; use_max_speed=false, ignore_unknown=false, modelname=modelname(MODELTYPE_DEFAULT, lang))
+function next_letter(lang::String; ignore_unknown=false)
+    if !ignore_unknown && is_next(UNKNOWN_TOKEN, [keys(ALPHABET[lang])...]; use_max_speed=false, ignore_unknown=false, modelname=modelname(MODELTYPE_DEFAULT, lang))
         return UNKNOWN_TOKEN
     else
         if     (lang == LANG.DE   ) next_letter_DE()
@@ -486,16 +486,16 @@ interpret_letters_EN_US(input::AbstractString) = (return ALPHABET[LANG.EN_US][in
 interpret_letters_ES(input::AbstractString)    = (return ALPHABET[LANG.ES   ][input])
 interpret_letters_FR(input::AbstractString)    = (return ALPHABET[LANG.FR   ][input])
 
-@voiceargs letter=>(model=MODELNAME.DEFAULT.DE,    valid_input=[keys(ALPHABET[LANG.DE   ])...], interpret_function=interpret_letters_DE,    ignore_unknown=false) next_letter_DE(letter::String)    = (return letter)
-@voiceargs letter=>(model=MODELNAME.DEFAULT.EN_US, valid_input=[keys(ALPHABET[LANG.EN_US])...], interpret_function=interpret_letters_EN_US, ignore_unknown=false) next_letter_EN_US(letter::String) = (return letter)
-@voiceargs letter=>(model=MODELNAME.DEFAULT.ES,    valid_input=[keys(ALPHABET[LANG.ES   ])...], interpret_function=interpret_letters_ES,    ignore_unknown=false) next_letter_ES(letter::String)    = (return letter)
-@voiceargs letter=>(model=MODELNAME.DEFAULT.FR,    valid_input=[keys(ALPHABET[LANG.FR   ])...], interpret_function=interpret_letters_FR,    ignore_unknown=false) next_letter_FR(letter::String)    = (return letter)
+@voiceargs letter=>(model=MODELNAME.DEFAULT.DE,    valid_input=[keys(ALPHABET[LANG.DE   ])...], interpret_function=interpret_letters_DE,    ignore_unknown=true) next_letter_DE(letter::String)    = (return letter)
+@voiceargs letter=>(model=MODELNAME.DEFAULT.EN_US, valid_input=[keys(ALPHABET[LANG.EN_US])...], interpret_function=interpret_letters_EN_US, ignore_unknown=true) next_letter_EN_US(letter::String) = (return letter)
+@voiceargs letter=>(model=MODELNAME.DEFAULT.ES,    valid_input=[keys(ALPHABET[LANG.ES   ])...], interpret_function=interpret_letters_ES,    ignore_unknown=true) next_letter_ES(letter::String)    = (return letter)
+@voiceargs letter=>(model=MODELNAME.DEFAULT.FR,    valid_input=[keys(ALPHABET[LANG.FR   ])...], interpret_function=interpret_letters_FR,    ignore_unknown=true) next_letter_FR(letter::String)    = (return letter)
 
 
 
 "Get next digit from speech."
[…]
