### 2022-12-08 276d0cb
force new command recognizer after recognizer change
