### 2022-12-08 58f15a8
force new command recognizer after recognizer change
