### 2022-12-08 3b27a2d
force new command recognizer after recognizer change
