### 2022-12-09 dba4fcd
add functions to type capitals
