### 2023-01-09 #121 @whedon
[REVIEW]: JustSayIt.jl: A Fresh Approach to Open Source Voice Assistant Development
**Submitting author:** @omlins (<a href="http://orcid.org/0000-0001-8969-5619">Samuel Omlin</a>)
**Repository:** https://github.com/omlins/JustSayIt.jl
**Branch with paper.md** (empty if default branch): JuliaConProceeding2022
**Version:** 0.3.0

- [ ] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?

- [x] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?
