### 2023-01-30 omlins/JustSayIt.jl#87 @pfitzseb
Usage example fails due to incorrect type checks
The example at https://omlins.github.io/JustSayIt.jl/stable/usage/ fails due to the types specified at https://github.com/omlins/JustSayIt.jl/blob/7ec7f1bdfe90975e3de485a7cd406889ecbd73a9/src/init_jsi.jl#L29 not including `typeof(Key.ctrl) == PyObject`.

Ref https://github.com/JuliaCon/proceedings-review/issues/121.
