### 2023-01-30 omlins/JustSayIt.jl#88 @pfitzseb
Keyboard.type fails with "insecure command argument recognition"
Following the usage instructions at https://omlins.github.io/JustSayIt.jl/stable/usage/#Fast-command-programming-with-voice-argument-functions, `type` fails:
```
julia> start(commands=commands, max_speed_subset=["ma", "select", "okay", "middle", "right", "double", "triple", "copy", "upwards", "downwards", "take"])
[ Info: JustSayIt: I am initializing...
[ Info: Listening for commands in English (United States) (say "sleep JustSayIt" to put me to sleep; press CTRL+c to terminate)...
[ Info: Starting command: click_left (latency: 11 ms)
[ Info: Starting command: type
[ Info: Command `type` aborted: insecure command argument recognition.
^C[ Info: Terminating JustSayIt...
[ Info: Stopped listening for commands.
```

Ref https://github.com/JuliaCon/proceedings-review/issues/121.
