### 2023-02-12 omlins/JustSayIt.jl#88 @omlins
@pfitzseb : thanks for the issue. However, I don't think the link you posted points points to the right example. Could you please point to the right one?

Then, it might be here rather a documentation issue then a software issue, i.e., the documentation could potentially be clearer. Concretely, after starting the command "type" you need to specify the type mode; else it will abort as shown in your snippet. To see the available type modes say "help type" (more info about help is found here: https://omlins.github.io/JustSayIt.jl/stable/usage/#Help-on-commands-callable-by-voice)
