### 2023-02-13 omlins/JustSayIt.jl#88 @pfitzseb
Oh, gotcha. Definitely user error, I didn't realize `type` takes a second argument for some reason. The error message isn't particularly helpful though :)
