+++ previews/PR89/examples.MD
@@ -1,0 +1,5 @@
+# Examples
+
+```@contents
+Pages = ["examples/config_custom_function.md"]
+```
