+++ .github/workflows/CI.yml
@@ -20,7 +20,7 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '1.7' # Minimum required Julia version.
+          - '1.6' # Minimum required Julia version.
           - '1'   # Latest stable 1.x release of Julia
           #- 'nightly'
         os:
