### 2023-04-17 9649a86
ensure that the submodules also use conda.jl miniforge
