+++ src/Commands/Keyboard.jl
@@ -26,7 +26,6 @@ const Pynput = PyNULL()
 function __init__()
     if !haskey(ENV, "JSI_USE_PYTHON") ENV["JSI_USE_PYTHON"] = "1" end
     if ENV["JSI_USE_PYTHON"] == "1"
-        ENV["PYTHON"] = ""                                              # Force PyCall to use Conda.jl
         copy!(Pynput, pyimport_pip("pynput"))
         set_controller("keyboard", Pynput.keyboard.Controller())
     end
