### 2023-04-17 1fc7b80
ensure that the submodules also use conda.jl miniforge
