+++ src/shared.jl
@@ -165,13 +165,13 @@ function force_miniforge()
     ENV["CONDA_JL_USE_MINIFORGE"] = "1"                             # Force usage of miniforge
     if !Conda.USE_MINIFORGE
         @info "Rebuilding Conda.jl for using Miniforge..."
-        run(Cmd(`$(Base.julia_cmd()) -e 'import Pkg; Pkg.build("Conda")'`, env=("CONDA_JL_USE_MINIFORGE" => "1",))) #Pkg.build("Conda")
+        run(Cmd(`$(Base.julia_cmd()) --project="$(Base.active_project())" -e 'import Pkg; Pkg.build("Conda")'`, env=("CONDA_JL_USE_MINIFORGE" => "1",))) #Pkg.build("Conda")
         do_restart = true
     end
     ENV["PYTHON"] = ""                                              # Force PyCall to use Conda.jl
     if !any(startswith.(PyCall.python, DEPOT_PATH))                 # Rebuild of PyCall if it has not been built with Conda.jl (alternative check could be !PyCall.conda, however this would set the version compatibility to when it was introduced)
         @info "Rebuilding PyCall for using Julia Conda.jl..."
-        run(Cmd(`$(Base.julia_cmd()) -e 'import Pkg; Pkg.build("PyCall")'`, env=("CONDA_JL_USE_MINIFORGE" => "1",))) #Pkg.build("PyCall")
+        run(Cmd(`$(Base.julia_cmd()) --project="$(Base.active_project())" -e 'import Pkg; Pkg.build("PyCall")'`, env=("CONDA_JL_USE_MINIFORGE" => "1",))) #Pkg.build("PyCall")
         do_restart = true
     end
     if do_restart
