### 2023-04-21 0c3e97b
ensure that the submodules also use conda.jl miniforge
