### 2023-04-21 8dc4c5d
update check whether PyCall is installed
