+++ previews/PR92/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"api/","page":"High-level API reference","title":"High-level API reference","text":"CurrentModule = JustSayIt","category":"page"},{"location":"api/#High-level-API-reference","page":"High-level API reference","title":"High-level API reference","text":"","category":"section"},{"location":"api/","page":"High-level API reference","title":"High-level API reference","text":"This is the official High-level API reference of JustSayIt. Note that it can also be queried interactively from the Julia REPL using the help mode:","category":"page"},{"location":"api/","page":"High-level API reference","title":"High-level API reference","text":"julia> using JustSayIt\njulia>?\nhelp?> JustSayIt.API","category":"page"},{"location":"api/#Submodule","page":"High-level API reference","title":"Submodule","text":"","category":"section"},{"location":"api/","page":"High-level API reference","title":"High-level API reference","text":"JustSayIt.API","category":"page"},{"location":"api/#JustSayIt.API","page":"High-level API reference","title":"JustSayIt.API","text":"Module API\n\nApplication Programming Interface (API) of JustSayIt.\n\nMacros\n\n@voiceargs\n\nFunctions\n\nis_next\nare_next\npause_recording\nrestart_recording\n\nConstants\n\nLANG\nMODELNAME\n\nTo see a description of a function, macro or module type ?<functionname>, ?<macroname> (including the @) or ?<modulename>, respectively.\n\n\n\n\n\n","category":"module"},{"location":"api/#Macros","page":"High-level API reference","title":"Macros","text":"","category":"section"},{"location":"api/#Index","page":"High-level API reference","title":"Index","text":"","category":"section"},{"location":"api/","page":"High-level API reference","title":"High-level API reference","text":"JustSayIt.API.@voiceargs","category":"page"},{"location":"api/#Documentation","page":"High-level API reference","title":"Documentation","text":"","category":"section"},{"location":"api/","page":"High-level API reference","title":"High-level API reference","text":"JustSayIt.API.@voiceargs","category":"page"},{"location":"api/#JustSayIt.@voiceargs","page":"High-level API reference","title":"JustSayIt.@voiceargs","text":"@voiceargs args function\n\nDeclare some or all arguments of the function definition to be arguments that can be obtained by voice.\n\nArguments\n\nargs: a voicearg, a pair voicearg=>kwargs or a tuple with multiple voiceargs, which can each have kwargs or not (see examples below).\nfunction: the function definition.\n\nnote: Keyword arguments definable for each voice argument in `args`\nmodel::String=MODELNAME.DEFAULT.<default_language>: the name of the model to be used for the function argument (the name must be one of the keys of the modeldirs dictionary passed to init_jsi).\nvalid_input::AbstractArray{String}|NTuple{N,Pair{String,<:AbstractArray{String}}}: the valid speech input (e.g. [\"up\", \"down\"] or (\"de\" => [\"rauf\", \"runter\"], \"en-us\" => [\"up\", \"down\"], \"fr\" => [\"haut\", \"bas\"], \"es\" => [\"arriba\", \"abajo\"])).\nvalid_input_auto::Bool: whether the valid speech input can automatically be derived from the type of the function argument.\ninterpret_function::Function: a function to interpret the token (mapping a String to a different String).\nuse_max_speed::Bool=false: whether to use maxium speed for the recognition of the next token (rather than maximum accuracy). It is generally only recommended to set use_max_speed=true for single word commands or very specfic use cases that require immediate minimal latency action when a command is said.\nignore_unknown::Bool=false: whether to ignore unknown tokens in the speech (and consume the next instead). It is generally not recommended to set ignore_unkown=true - in particular not in combination with a limited valid input - as the
[…]
