### 2023-04-21 #121 @omlins
@carstenbauer : I had addressed some of the issues only partially until recently as the second review is still pending. However, I have now been able to address all the issues raised by @pfitzseb conclusively, with exception of the last one.

Here, @pfitzseb and @carstenbauer, you find a reply to each issue raised:

>     * In Fig 2, a `using JustSayIt.API` is required (same goes for the documentation at https://omlins.github.io/JustSayIt.jl/stable/usage/

Some stale commands in the `max_speed_subset` had caused the error when executing the code in Fig 2 (resolved here: https://github.com/omlins/JustSayIt.jl/pull/91). This error does not seem to be present in the documentation.

>     * [Usage example fails due to incorrect type checks omlins/JustSayIt.jl#87](https://github.com/omlins/JustSayIt.jl/issues/87) causes the example in Fig 2 to be non-functional.

This code issue has been solved and the corresponding pull request merged (https://github.com/omlins/JustSayIt.jl/pull/89).

>     * [Keyboard.type fails with "insecure command argument recognition" omlins/JustSayIt.jl#88](https://github.com/omlins/JustSayIt.jl/issues/88) causes the `Keyboard.type` command to be non-functional.

This was not an issue with the code, but an incorrect usage of JustSayIt (https://github.com/omlins/JustSayIt.jl/issues/88#issuecomment-1427649211).

>     * I'm seeing worse latency than shown in the article, but I'm happy to chalk that up to benchmarking on a signigicantly less powerful machine.

I believe it is right to expect that the latency will be different from system to system.

>     * There are no community guidelines. I'd recommend linking something like https://github.com/SciML/ColPrac

The contribution guide line has been extended and a link to the mentioned guideline added (https://github.com/omlins/JustSayIt.jl/pull/92).

>     * Some comparisons to other implementations of a text-to-command software would be nice (in terms of API, performance, latency)

I would prefer to wait for the second review to address this issue, because additions to the text mean also removal of existing parts, given the very short format of this paper (we are already at the limit of the two pages).

In addition to these issues, I've also been able to solve the longstanding issue with the automatic installation not working in some corner cases, which @pfitzseb observed as well (resolved here: https://github.com/omlins/JustSayIt.jl/pull/90). Note that independent of all the issues raised, two unit tests fail since recently on some versions of Ubuntu (the unit testing is complex, because the Vosk speech recognition tool kit used is not fully deterministic). They are not critical and could be removed in another pull request.
