### 2023-04-27 f03d1ec
enable controlling use of static versus dynamic recognizers in voiceargs
