### 2023-04-27 620c634
check if recognized token invalid input; add debug output
