### 2023-04-27 a8c2070
add valid tokens field to Recognizer
