### 2023-04-27 f0be593
remove force reset in start
