+++ src/start.jl
@@ -208,8 +208,7 @@ function start(; default_language::String=LANG.EN_US, type_languages::Union{Stri
                 end
             end
             try
-                if !is_active(recognizer(COMMAND_RECOGNIZER_ID)) update_commands() end
-                force_reset_previous(recognizer(COMMAND_RECOGNIZER_ID))
+                if !is_active(recognizer(COMMAND_RECOGNIZER_ID)) update_commands() end # NOTE: a reset might be beneficial or needed at some point, as previously force_reset_previous(recognizer(COMMAND_RECOGNIZER_ID))
                 use_max_speed = _is_next(max_speed_token_subset, recognizer(COMMAND_RECOGNIZER_ID), _noises(modelname_default); use_partial_recognitions=true, ignore_unknown=false)
                 cmd_name = next_token(recognizer(COMMAND_RECOGNIZER_ID), _noises(modelname_default); use_partial_recognitions = use_max_speed, ignore_unknown=false)
                 if cmd_name == UNKNOWN_TOKEN # For increased recognition security, ignore the current word group if the unknown token was obtained as command name (achieved by doing a full reset). This will prevent for example "text right" or "text type text" to trigger an action, while "right" or "type text" does so.
