### 2023-04-27 2b9e128
prevent command interpretation after exit on unknown from typing letters
