### 2023-04-28 05359bd
make token buffer comparisons space insensitive
