+++ src/Commands/Keyboard.jl
@@ -16,18 +16,22 @@ module Keyboard
 
 using PyCall
 using ..Exceptions
-import ..JustSayIt: @voiceargs, pyimport_pip, controller, set_controller, PyKey, default_language, type_languages, lang_str, LANG, LANG_CODES_SHORT, LANG_STR, ALPHABET, DIGITS, MODELTYPE_DEFAULT, MODELNAME, modelname, tic, toc, is_next, are_next, all_consumed, was_partial_recognition, InsecureRecognitionException, reset_all, do_delayed_resets, interpret_enum, UNKNOWN_TOKEN
+import ..JustSayIt: @voiceargs, pyimport_pip, controller, set_controller, PyKey, default_language, type_languages, lang_str, LANG, LANG_CODES_SHORT, LANG_STR, ALPHABET, DIGITS, DIRECTIONS, SIDES, COUNTS, SYMBOLS, MODELTYPE_DEFAULT, MODELNAME, modelname, tic, toc, is_next, are_next, next_token, next_tokengroup, all_consumed, was_partial_recognition, InsecureRecognitionException, reset_all, do_delayed_resets, interpret_enum, UNKNOWN_TOKEN
 
 
 ## PYTHON MODULES
 
-const Pynput = PyNULL()
+const Pynput    = PyNULL()
+const Tkinter   = PyNULL()
+# const Pyperclip = PyNULL()
 
 function __init__()
     if !haskey(ENV, "JSI_USE_PYTHON") ENV["JSI_USE_PYTHON"] = "1" end
     if ENV["JSI_USE_PYTHON"] == "1"
         ENV["PYTHON"] = ""                                              # Force PyCall to use Conda.jl
-        copy!(Pynput, pyimport_pip("pynput"))
+        copy!(Pynput,    pyimport_pip("pynput"))
+        copy!(Tkinter,   pyimport_pip("tkinter"))
+        # copy!(Pyperclip, pyimport_pip("pyperclip"))
         set_controller("keyboard", Pynput.keyboard.Controller())
     end
 end
@@ -131,6 +135,13 @@ const TYPE_KEYWORDS = Dict(
                        "exclamation"   => "exclamation",
                        "interrogation" => "interrogation",
                        "paragraph"     => "paragraphe"),
+)
+
+const WHITESPACE_PATTERN = Dict(
+    LANG.DE    => Dict("leer"    => r"\s+"),
+    LANG.EN_US => Dict("space"   => r"\s+"),
+    LANG.ES    => Dict("espacio" => r"\s+"),
+    LANG.FR    => Dict("espace"  => r"\s+"),
 )
 
 
@@ -469,14 +480,14 @@ end
 
 
 "Get next letter from speech."
-function next_letter(lang::String; ignore_unknown=false)
-    if !ignore_unknown && is_next(UNKNOWN_TOKEN, [keys(ALPHABET[lang])...]; use_max_speed=false, ignore_unknown=false, modelname=modelname(MODELTYPE_DEFAULT, lang))
+function next_letter(lang::String; ignore_unknown=false, use_max_speed=false)
+    if !ignore_unknown && is_next(UNKNOWN_TOKEN, [keys(ALPHABET[lang])...]; use_max_speed=use_max_speed, ignore_unknown=false, modelname=modelname(MODELTYPE_DEFAULT, lang))
         return UNKNOWN_TOKEN
     else
-        if     (lang == LANG.DE   ) next_letter_DE()
-        elseif (lang == LANG.EN_US) next_letter_EN_US()
-        elseif (lang == LANG.ES   ) next_letter_ES()
-        elseif (lang == LANG.FR   ) next_letter_FR()
[…]
