### 2023-05-18 550a647
add function next_tokengroup and dictionaries
