### 2023-05-18 41f770d
make token buffer comparisons space insensitive
