+++ src/shared.jl
@@ -175,6 +175,249 @@ const DIGITS = Dict(
     LANG.ES    => Dict("cero"=>"0", "uno"=>"1",  "duo"=>"2",  "tres"=>"3",  "quatro"=>"4", "cinco"=>"5",  "seis"=>"6",  "siete"=>"7",  "ocho"=>"8",  "nueve"=>"9", "punto"=>".", "coma"=>",",    "espacio"=>" "),
     LANG.FR    => Dict("zéro"=>"0", "un"=>"1",   "deux"=>"2", "trois"=>"3", "quatre"=>"4", "cinque"=>"5", "six"=>"6",   "sept"=>"7",   "huit"=>"8",  "neuf"=>"9",  "point"=>".", "virgule"=>",", "espace"=>" "),
 )
+const DIRECTIONS = Dict(
+    LANG.DE    => Dict("rechts"=>"right", "links"=>"left", "oben"=>"up", "unten"=>"down", "vorwärts"=>"forward", "rückwärts"=>"backward", "aufwärts"=>"upward", "abwärts"=>"downward"),
+    LANG.EN_US => Dict("right"=>"right", "left"=>"left", "up"=>"up",   "down"=>"down", "forward"=>"forward", "backward"=>"backward", "upward"=>"upward", "downward"=>"downward"),
+    LANG.ES    => Dict("derecha"=>"right", "izquierda"=>"left", "arriba"=>"up", "abajo"=>"down", "adelante"=>"forward", "atrás"=>"backward", "subiendo"=>"upward", "bajando"=>"downward"),
+    LANG.FR    => Dict("droite"=>"right", "gauche"=>"left", "haut"=>"up", "bas"=>"down", "avant"=>"forward", "arrière"=>"backward", "montant"=>"upward", "descendant"=>"downward"),
+)
+const SIDES = Dict(
+    LANG.DE    => Dict("vor"=>"before", "nach"=>"after"),
+    LANG.EN_US => Dict("before"=>"before", "after"=>"after"),
+    LANG.ES    => Dict("antes"=>"before", "después"=>"after"),
+    LANG.FR    => Dict("avant"=>"before", "après"=>"after"),
+)
+const COUNTS = Dict(
+    LANG.DE    => Dict("eins"=>"1", "zwei"=>"2", "drei"=>"3", "vier"=>"4", "fünf"=>"5", "sechs"=>"6", "sieben"=>"7", "acht"=>"8", "neun"=>"9", "zehn"=>"10", "fünfzig"=>"50", "hundert"=>"100", "tausend"=>"1000"),
+    LANG.EN_US => Dict("one"=>"1", "two"=>"2", "three"=>"3", "four"=>"4", "five"=>"5", "six"=>"6", "seven"=>"7", "eight"=>"8", "nine"=>"9", "ten"=>"10", "fifty"=>"50", "hundred"=>"100", "thousand"=>"1000"),
+    LANG.ES    => Dict("uno"=>"1", "dos"=>"2", "tres"=>"3", "cuatro"=>"4", "cinco"=>"5", "seis"=>"6", "siete"=>"7", "ocho"=>"8", "nueve"=>"9", "diez"=>"10", "cincuenta"=>"50", "cien"=>"100", "mil"=>"1000"),
+    LANG.FR    => Dict("un"=>"1", "deux"=>"2", "trois"=>"3", "quatre"=>"4", "cinq"=>"5", "six"=>"6", "sept"=>"7", "huit"=>"8", "neuf"=>"9", "dix"=>"10", "cinquante"=>"50", "cent"=>"100", "mille"=>"1000"),
+)
+
+
+# Sign commands.
+
+PUNCTUATION_SYMBOLS = Dict(
+    LANG.DE    => Dict("punkt"                       => ".",
[…]
