### 2023-05-18 48dcb66
enable ignore_unknown and use_max_speed as keyword arguments in voicearg functions
