### 2023-05-18 43fee8f
split navigate function
