+++ src/Commands/Keyboard.jl
@@ -642,124 +642,141 @@ function type_fixedcase(case::Case)
 end
 
 
-function navigate(; up_lines=10, down_lines=10)
-    Key        = Pynput.keyboard.Key
+function navigate(; up_lines=10, down_lines=10, action="go")
     lang       = default_language()
     directions = [keys(DIRECTIONS[lang])...]
     counts     = [keys(COUNTS[lang])...]
     modes      = [directions..., counts...]
+    token      = next_token(modes; consume=false, use_partial_recognitions=true, ignore_unknown=false)        # NOTE: the token is not consumed yet...
+    if     (token in counts)     move_cursor_simple(; action=action)                                          # Case: go <count> <direction>
+    elseif (token in directions) move_cursor_smart(; up_lines=up_lines, down_lines=down_lines, action=action) # Case: go <direction> <side> <text>
+    else                         @InsecureRecognitionException("unknown direction or count.")                 # NOTE: it could return without exception, if the command handler will just continue from there.
+    end
+end
 
-    token = next_token(modes; consume=false, use_partial_recognitions=true, ignore_unknown=false) # NOTE: the token is not consumed yet...
 
-    # Case: go <direction> <side> <text>
-    if (token in directions)
-        direction = next_direction(; use_max_speed=true) #TODO: see if forward backward upward and downward should be included here.
-        if (token != UNKNOWN_TOKEN)        
-            @info "go $direction <side> <text> ..."
-            @show context = get_context(direction, up_lines, down_lines)
-            @show valid_input = derive_valid_input(context) # NOTE: pre compute the valid input while the side key word is being spoken.
-            token = next_side(; use_max_speed=true)
-            if (token != UNKNOWN_TOKEN)
-                side = token
-                @info "... go $direction $side <text> ..."
-                @show text_keys = next_tokengroup(valid_input; use_partial_recognitions=false, ignore_unknown=false)
-                if !all(text_keys .== UNKNOWN_TOKEN)
-                    @show text_elements = replace(text_keys, SYMBOLS[lang]..., WHITESPACE_PATTERN[lang]..., UNKNOWN_TOKEN => r"[\\p{L}\\p{S}]") # NOTE: \p{L} = letters, \p{N} = numbers, \p{P} = punctuation, \p{S} = symbols, \p{Z} = whitespace, \p{M} = marks, \p{C} = other characters
-                    @show text_pattern = prod(regex.(text_elements, "is")) # NOTE: the text elements also include spaces, so we need to join without spaces.
-                    @show range = findfirst(text_pattern, context)
-                    if isnothing(range) # NOTE: letter recognition is not very secure; thus, as no match had been found, we replace letters with a pattern that matches any letter.
-                        @show text_elements = replace(text_elements, r"[a-z]"  => r"[a-z]")
-                        @show text_pattern = prod(regex.(text_elements, "is")) # NOTE: the text elements also include spaces, so we need to join without spaces.
-                        @show range = findfirst(text_pattern, context)
-                    end
-                    if !isnothing(range)
-                        @info "... go $direction $side $text_pattern"
-                        stop_highlighting(direction; to_origin=false)
-                        @show displacement_from_origin = move_cursor(side, direction, context, range)
-                        token = next_direction(; use_max_speed=true) #TODO: see if other then forward backward should be included here.
-                        ranges = [range]
-                        was_forward = true
-                        while token in ["forward", "backward"]
-                            if token == "forward"  
-                                @show range = findnext(text_pattern, context, ranges[end].stop+1)
-                                if isnothing(range)
-                                    @info "... no more match found."
-                                else
-                                    @info "... forward ..."
-                                    push!(ranges, range)
-                                end
-                                was_forward = true
[…]
