+++ src/Commands/Keyboard.jl
@@ -144,6 +144,13 @@ const WHITESPACE_PATTERN = Dict(
     LANG.FR    => Dict("espace"  => r"\s+"),
 )
 
+const TEXTRANGE_PATTERN = Dict(
+    LANG.DE    => Dict("bis"     => r".+"is),
+    LANG.EN_US => Dict("till"    => r".+"is),
+    LANG.ES    => Dict("hasta"   => r".+"is),
+    LANG.FR    => Dict("jusqu'à" => r".+"is),
+)
+
 
 ## FUNCTIONS
 
@@ -642,21 +649,53 @@ function type_fixedcase(case::Case)
 end
 
 
-function navigate(; up_lines=10, down_lines=10, action="go")
+function navigate_smart(; up_lines=10, down_lines=10, action="go", select=false, copy=false, select_prefix=(Pynput.keyboard.Key.shift,), copy_shortcut=(Pynput.keyboard.Key.ctrl, 'c'))
+    if (copy) select = true end
     lang       = default_language()
     directions = [keys(DIRECTIONS[lang])...]
     counts     = [keys(COUNTS[lang])...]
     modes      = [directions..., counts...]
     token      = next_token(modes; consume=false, use_partial_recognitions=true, ignore_unknown=false)        # NOTE: the token is not consumed yet...
-    if     (token in counts)     move_cursor_simple(; action=action)                                          # Case: go <count> <direction>
-    elseif (token in directions) move_cursor_smart(; up_lines=up_lines, down_lines=down_lines, action=action) # Case: go <direction> <side> <text>
-    else                         @InsecureRecognitionException("unknown direction or count.")                 # NOTE: it could return without exception, if the command handler will just continue from there.
+    if token in counts               # Case: <action> <count> <direction>
+        prefix = select ? select_prefix : ()
+        move_cursor(; prefix=prefix, action=action)
+        if (copy) press_keys(copy_shortcut...) end
+    elseif token in directions       # Case: <action> <direction> <text localization>
+        move_cursor_smart(; up_lines=up_lines, down_lines=down_lines, action=action, select=select, copy=copy, select_prefix=select_prefix, copy_shortcut=copy_shortcut)
+    else
+        @InsecureRecognitionException("unknown direction or count.") # NOTE: it could return without exception, if the command handler will just continue from there.
+    end
+end
+
+function select_smart(; up_lines=10, down_lines=10, action="light", select_prefix=(Pynput.keyboard.Key.shift,))
+    navigate_smart(; up_lines=up_lines, down_lines=down_lines, action=action, select=true, select_prefix=select_prefix)
+end
+
+function copy_smart(; up_lines=10, down_lines=10, action="take", select_prefix=(Pynput.keyboard.Key.shift,), copy_shortcut=(Pynput.keyboard.Key.ctrl, 'c'))
+    navigate_smart(; up_lines=up_lines, down_lines=down_lines, action=action, copy=true, select_prefix=select_prefix, copy_shortcut=copy_shortcut)
+end
+
+
+function move_cursor(count::Integer, direction::String; prefix=())
+    Key = Pynput.keyboard.Key
+    if     (direction == "right")    press_keys(prefix..., Key.right;           count=count)
+    elseif (direction == "left")     press_keys(prefix..., Key.left;            count=count)
+    elseif (direction == "up")       press_keys(prefix..., Key.up;              count=count)
+    elseif (direction == "down")     press_keys(prefix..., Key.down;            count=count)
+    elseif (direction == "forward")  press_keys(prefix..., Key.ctrl, Key.right; count=count)
+    elseif (direction == "backward") press_keys(prefix..., Key.ctrl, Key.left;  count=count)
[…]
