### 2023-05-22 06056ee
introduce text regions and remove need of restoring clipboard
