+++ src/Commands/Keyboard.jl
@@ -16,7 +16,7 @@ module Keyboard
 
 using PyCall
 using ..Exceptions
-import ..JustSayIt: @voiceargs, pyimport_pip, controller, set_controller, PyKey, default_language, type_languages, lang_str, LANG, LANG_CODES_SHORT, LANG_STR, ALPHABET, DIGITS, DIRECTIONS, SIDES, COUNTS, SYMBOLS, MODELTYPE_DEFAULT, MODELNAME, modelname, tic, toc, is_next, are_next, next_token, next_tokengroup, all_consumed, was_partial_recognition, InsecureRecognitionException, reset_all, do_delayed_resets, interpret_enum, UNKNOWN_TOKEN
+import ..JustSayIt: @voiceargs, pyimport_pip, controller, set_controller, PyKey, default_language, type_languages, lang_str, LANG, LANG_CODES_SHORT, LANG_STR, ALPHABET, DIGITS, DIRECTIONS, REGIONS, SIDES, COUNTS, SYMBOLS, MODELTYPE_DEFAULT, MODELNAME, modelname, tic, toc, is_next, are_next, next_token, next_tokengroup, all_consumed, was_partial_recognition, InsecureRecognitionException, reset_all, do_delayed_resets, interpret_enum, UNKNOWN_TOKEN
 
 
 ## PYTHON MODULES
@@ -145,10 +145,10 @@ const WHITESPACE_PATTERN = Dict(
 )
 
 const TEXTRANGE_PATTERN = Dict(
-    LANG.DE    => Dict("bis"     => r".+"is),
-    LANG.EN_US => Dict("till"    => r".+"is),
-    LANG.ES    => Dict("hasta"   => r".+"is),
-    LANG.FR    => Dict("jusqu'à" => r".+"is),
+    LANG.DE    => Dict("bis"     => r".+?"is),
+    LANG.EN_US => Dict("till"    => r".+?"is),
+    LANG.ES    => Dict("hasta"   => r".+?"is),
+    LANG.FR    => Dict("jusqu'à" => r".+?"is),
 )
 
 
@@ -654,13 +654,14 @@ function navigate_smart(; up_lines=10, down_lines=10, action="go", select=false,
     lang       = default_language()
     directions = [keys(DIRECTIONS[lang])...]
     counts     = [keys(COUNTS[lang])...]
-    modes      = [directions..., counts...]
+    regions    = [keys(REGIONS[lang])...]
+    modes      = [directions..., counts..., regions...]
     token      = next_token(modes; consume=false, use_partial_recognitions=true, ignore_unknown=false)        # NOTE: the token is not consumed yet...
-    if token in counts               # Case: <action> <count> <direction>
+    if token in counts                              # Case: <action> <count> <direction>
         prefix = select ? select_prefix : ()
         move_cursor(; prefix=prefix, action=action)
         if (copy) press_keys(copy_shortcut...) end
-    elseif token in directions       # Case: <action> <direction> <text localization>
+    elseif token in regions                         # Case: <action> <region> <text range>
         move_cursor_smart(; up_lines=up_lines, down_lines=down_lines, action=action, select=select, copy=copy, select_prefix=select_prefix, copy_shortcut=copy_shortcut)
     else
         @InsecureRecognitionException("unknown direction or count.") # NOTE: it could return without exception, if the command handler will just continue from there.
@@ -720,40 +721,46 @@ end
 function move_cursor_smart(; up_lines=10, down_lines=10, action="go", select=false, copy=false, select_prefix=(Pynput.keyboard.Key.shift,), copy_shortcut=(Pynput.keyboard.Key.ctrl, 'c'))
     if (copy) select = true end
     lang              = default_language()
[…]
