### 2023-06-23 b335d19
add resolved navigation in smart move
