+++ src/Commands/Keyboard.jl
@@ -16,7 +16,7 @@ module Keyboard
 
 using PyCall
 using ..Exceptions
-import ..JustSayIt: @voiceargs, pyimport_pip, controller, set_controller, PyKey, default_language, type_languages, lang_str, LANG, LANG_CODES_SHORT, LANG_STR, ALPHABET, DIGITS, DIRECTIONS, REGIONS, SIDES, COUNTS, SYMBOLS, MODELTYPE_DEFAULT, MODELNAME, modelname, tic, toc, is_next, are_next, next_token, next_tokengroup, all_consumed, was_partial_recognition, InsecureRecognitionException, reset_all, do_delayed_resets, interpret_enum, UNKNOWN_TOKEN
+import ..JustSayIt: @voiceargs, pyimport_pip, controller, set_controller, PyKey, default_language, type_languages, lang_str, LANG, LANG_CODES_SHORT, LANG_STR, ALPHABET, DIGITS, DIRECTIONS, REGIONS, FRAGMENTS, SIDES, COUNTS, SYMBOLS, MODELTYPE_DEFAULT, MODELNAME, modelname, tic, toc, is_next, are_next, next_token, next_tokengroup, all_consumed, was_partial_recognition, InsecureRecognitionException, reset_all, do_delayed_resets, interpret_enum, UNKNOWN_TOKEN
 
 
 ## PYTHON MODULES
@@ -149,6 +149,13 @@ const TEXTRANGE_PATTERN = Dict(
     LANG.EN_US => Dict("till"    => r".+?"is),
     LANG.ES    => Dict("hasta"   => r".+?"is),
     LANG.FR    => Dict("jusqu'à" => r".+?"is),
+)
+
+const NAVIGATION_ACTIONS = Dict(
+    LANG.DE    => Dict("vorwärts" => "forward", "rückwärts" => "backward", "einfügen" => "paste", "löschen"   => "delete", "bis"     => "till", "beenden" => "exit"),
+    LANG.EN_US => Dict("forward"  => "forward", "backward"  => "backward", "paste"    => "paste", "delete"    => "delete", "till"    => "till", "exit"    => "exit"),
+    LANG.ES    => Dict("adelante" => "forward", "atrás"     => "backward", "pegar"    => "paste", "borrar"    => "delete", "hasta"   => "till", "salir"   => "exit"),
+    LANG.FR    => Dict("avancer"  => "forward", "reculer"   => "backward", "coller"   => "paste", "supprimer" => "delete", "jusqu'à" => "till", "sortir"  => "exit"),
 )
 
 
@@ -561,6 +568,13 @@ function type_string(str::String; do_keystrokes::Bool=true)
     if do_keystrokes
         keyboard = controller("keyboard")
         keyboard.type(str)
+        # NOTE: the following is an alternative for typing strings, which could be either more or less robust on certain systems:
+        # keys = map(convert_key, [str...])
+        # for key in keys
+        #     keyboard.press(key)
+        #     keyboard.release(key)
+        # end
+
     end
 end
 
@@ -635,6 +649,19 @@ type_camelcase() = type_fixedcase(camel)
 "Type words in constantcase (uppercase, separated with underscore)."
 type_constantcase() = type_fixedcase(constant)
 
+"Type digit string (exit on unknown)."
+function type_digit_string()
+    token = next_digit(default_language()) # NOTE: one token has to be queried in any case to be sure that we are inside a word group.
+    if (token != UNKNOWN_TOKEN)
+        digit_string = token
+        while (token != UNKNOWN_TOKEN && !all_consumed()) # NOTE: we exit on unknown or at the end of the word group.
+            token = next_digit(default_language())
[…]
