### 2023-06-23 401a42f
remove stale clipboard code and comments
