### 2023-06-23 1db2659
temporarily deactivate subset check and move execute to shared
