### 2023-06-23 f18f196
temporarily add the keys of the frequent navigate commands to the valid input of the command recognizer
