+++ src/JustSayIt.jl
@@ -49,6 +49,7 @@ include("Commands/Help.jl")
 
 ## Alphabetical include of command-submodules (must be at end as needs to import from JustSayIt, .e.g. next_recognition, next_partial_recognition)
 include("Commands/Email.jl")
+include("Commands/Generic.jl")
 include("Commands/Internet.jl")
 
 ## Include of main application and API (must be at end as needs to import potentially anything available in JustSayIt, in particular the Commands submodules).
@@ -57,7 +58,7 @@ include("API.jl")
 
 ## Exports (need to be after include of submodules if re-exports from them)
 export start, @include_config
-export Keyboard, Mouse, Help, Email, Internet
+export Keyboard, Mouse, Help, Email, Generic, Internet
 export Key
 
 end # module JustSayIt
