### 2023-07-17 8a60d52
fixed keyboard python module init
