### 2023-07-17 ec0bc98
fix type initialization
