### 2023-07-17 de2e042
fix type initialization
