### 2023-07-17 a5ace33
fix type initialization
