### 2023-07-17 95316e0
update unit tests checking recognizer
