### 2023-07-21 2dd52d4
make unit tests compatible with Julia 1.9
