+++ versions.js
@@ -1,8 +1,9 @@
 var DOC_VERSIONS = [
   "stable",
+  "v0.3",
   "v0.2",
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.2.1";
+var DOCUMENTER_NEWEST = "v0.3.0";
 var DOCUMENTER_STABLE = "stable";
