+++ v0.3.0/usage.MD
@@ -1,0 +1,152 @@
+# Usage
+
+## Fully automatic installation
+After installing [Julia] - if not yet installed - JustSayIt can be installed directly with the [Julia package manager] from the [Julia REPL]:
+```julia-repl
+julia>]
+  pkg> add JustSayIt
+```
+All dependencies are automatically installed. Python dependencies are automatically installed in a local Miniconda environment, set up by [Conda.jl]. Default language models are automatically downloaded at first usage of JustSayIt.
+
+
+## Simple basic software usage
+The basic usage of JustSayIt is as simple as launching the [Julia REPL] and calling the commands `using JustSayIt` followed by `start()`:
+```julia-repl
+$> julia
+julia> using JustSayIt
+julia> start()
+```
+More information on customization keywords of [`start`](@ref) is given in the following subsections and in the [Software reference](@ref).
+
+
+## User definable mapping of command names to functions or keyboard shortcuts
+The keyword `commands` of [`start`](@ref) enables to freely define a mapping of command names to functions or keyboard shortcuts or sequences of those, e.g.:
+```julia
+# Define custom commands
+using JustSayIt
+commands = Dict("help"      => Help.help,
+                "type"      => Keyboard.type,
+                "ma"        => Mouse.click_left,
+                "select"    => Mouse.press_left,
+                "okay"      => Mouse.release_left,
+                "middle"    => Mouse.click_middle,
+                "right"     => Mouse.click_right,
+                "double"    => Mouse.click_double,
+                "triple"    => Mouse.click_triple,
+                "copy"      => (Key.ctrl, 'c'),
+                "cut"       => (Key.ctrl, 'x'),
+                "paste"     => (Key.ctrl, 'v'),
+                "undo"      => (Key.ctrl, 'z'),
+                "redo"      => (Key.ctrl, Key.shift, 'z'),
+                "upwards"   => Key.page_up,
+                "downwards" => Key.page_down,
+                "take"      => [Mouse.click_double, (Key.ctrl, 'c')],
+                "replace"   => [Mouse.click_double, (Key.ctrl, 'v')],
+                );
+start(commands=commands, max_speed_subset=["ma", "select", "okay", "middle", "right", "double", "triple", "copy", "upwards", "downwards", "take"])
+```
+
+Note that keyboard shortcuts are given as either single keys (e.g., `Key.page_up`) or tuples of keys (e.g., `(Key.ctrl, 'c')`). Special keys are selected from those available in `Key`: type `Key.`+ `tab` in the [Julia REPL] to see the available keys (the full documentation on the available keys is available [here](https://pynput.readthedocs.io/en/latest/keyboard.html#pynput.keyboard.Key)). Character keys are simply given in single quotes (e.g., `'c'`).
+
+A mapping of a command name to a sequence of commands can be defined by using as mapping target an array whose elements are functions or keyboard shortcuts (e.g., `[Mouse.click_double, (Key.ctrl, 'c')]`). This enables to express new complex actions without programming. In the above example, "take" double clicks the word below the mouse curser to select it and then copies it; "replace" double clicks the word below the mouse curser to select it and then replaces it.
+ 
+The keyword `subset` of [`start`](@ref) enables to activate only a subset of the default or user-defined commands. The following example selects a subset of the default commands:
+```julia
+# Listen only to the commands "help" and "type".
+using JustSayIt
+start(subset=["help", "type"])
+```
+
+
+## Per command choice between maximum speed or accuracy
+
+The keyword `max_speed_subset` of [`start`](@ref) enables to define a subset of the `commands` for which the command names are to be recognized with maximum speed rather than with maximum accuracy, e.g.:
+```julia
+# Define custom commands
+using JustSayIt
[…]
