+++ README.md
@@ -5,7 +5,13 @@
 [![CI](https://github.com/omlins/JustSayIt.jl/actions/workflows/CI.yml/badge.svg?branch=main)](https://github.com/omlins/JustSayIt.jl/actions/workflows/CI.yml?query=branch%3Amain)
 [![Coverage](https://codecov.io/gh/omlins/JustSayIt.jl/branch/main/graph/badge.svg)](https://codecov.io/gh/omlins/JustSayIt.jl)
 
-> :warning: JustSayIt was presented in a talk at JuliaCon 2022: [video of the talk](https://www.youtube.com/watch?v=W7oQb7pLc04), [abstract](https://pretalx.com/juliacon-2022/talk/H3N8UN) (the demo video mentioned in the talk was postponed to the next JuliaCon conference)
+> :warning: JustSayIt was first presented in a talk at JuliaCon 2022: [video of the talk](https://www.youtube.com/watch?v=W7oQb7pLc04), [abstract](https://pretalx.com/juliacon-2022/talk/H3N8UN)
+
+> :warning: A second talk "Quick Assembly of Personalized Voice Assistants with JustSayIt" was given at JuliaCon 2023: [video of the talk](https://www.youtube.com/watch?v=_gpH-mkrdGM), [abstract](https://pretalx.com/juliacon2023/talk/review/9MJFPDJV9DR7ANUXPSP9ZWJRFWSE83EY)
+
+> :warning: Contribute to [JustSayItRecipes.jl](https://github.com/omlins/JustSayItRecipes.jl), the official platform for sharing command dictionaries defining the assembly of personalized voice assistants using JustSayIt.jl! 
+
+
 
 This is the source repository of the Julia package `JustSayIt.jl`. Please head to the [documentation website](https://omlins.github.io/JustSayIt.jl) to learn about its usage:
 [![Preview](docs/preview.png)](https://omlins.github.io/JustSayIt.jl/)
