### 2024-10-22 #121 @pfitzseb
Alright, this ticks all the boxes now :)
