### 2024-11-05 #121 @omlins
> @omlins can you add the version and Zenodo DOI archive?

I just saw the notification. Yes, I will do so.
