### 2024-11-08 #121 @omlins
@matbesancon : here you can find the DOI for the paper: https://doi.org/10.5281/zenodo.14057465
here you can find the DOI for the code: https://doi.org/10.5281/zenodo.14057584
