### 2024-12-14 #121 @matbesancon
@omlins please check the paper before the final acceptance!
